// src/service-api/generated/models/getCompaniesSortBy.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * sales-crm
 * OpenAPI spec version: 1.0.0
 */

export type GetCompaniesSortBy = typeof GetCompaniesSortBy[keyof typeof GetCompaniesSortBy];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetCompaniesSortBy = {
  createdAt: 'createdAt',
  name: 'name',
  industry: 'industry',
} as const;