// src/service-api/generated/models/getCompaniesSortOrder.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * sales-crm
 * OpenAPI spec version: 1.0.0
 */

export type GetCompaniesSortOrder = typeof GetCompaniesSortOrder[keyof typeof GetCompaniesSortOrder];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetCompaniesSortOrder = {
  asc: 'asc',
  desc: 'desc',
} as const;